Validation

+++ src/main/java/com/caiofabio/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.caiofabio.dscommerce.services;
 import com.caiofabio.dscommerce.dto.ProductDTO;
 import com.caiofabio.dscommerce.entities.Product;
 import com.caiofabio.dscommerce.repositories.ProductRepository;
+import com.caiofabio.dscommerce.services.exceptions.DatabaseException;
 import com.caiofabio.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -52,21 +56,31 @@ public class ProductService {
     // atualizar pelo id
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-
-        Product entity = repository.getReferenceById(id);
-        //o metodo foi criado embaixo, boas praticas
-        copyDtoEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            //o metodo foi criado embaixo, boas praticas
+            copyDtoEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-
+    @Transactional(propagation = Propagation.SUPPORTS)
     // deletar pelo id
-    @Transactional
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
-
 
     // metodo craido para cadrasto e atualizacao
     private void copyDtoEntity(ProductDTO dto, Product entity) {
